pyproject.toml
[project]
name = "simplepromptmanage"
description = "Make Prompting as Easy as Grocery Shopping! "
version = "1.0.1"
license = {file = "LICENSE"} 
# classifiers = [
#     # For OS-independent nodes (works on all operating systems)
#     "Operating System :: OS Independent",
#
#     # OR for OS-specific nodes, specify the supported systems:
#     "Operating System :: Microsoft :: Windows",  # Windows specific
#     "Operating System :: POSIX :: Linux",  # Linux specific
#     "Operating System :: MacOS",  # macOS specific
#
#     # GPU Accelerator support. Pick the ones that are supported by your extension.
#     "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
#     "Environment :: GPU :: AMD ROCm",       # AMD ROCm support
#     "Environment :: GPU :: Intel Arc",      # Intel Arc support
#     "Environment :: NPU :: Huawei Ascend",  # Huawei Ascend support
#     "Environment :: GPU :: Apple Metal",    # Apple Metal support
# ]

dependencies = ["# ComfyUI Prompt Manager Requirements", "# Core dependencies", "aiohttp>=3.9.0", "Pillow>=10.0.0", "requests>=2.31.0", "# Optional: Performance optimization", "orjson>=3.9.0", "# Optional: Lora update service", "aiofiles>=23.0.0", "# Optional: Download from CivitAI (for download_by_civitaiwebnum.py)", "selenium>=4.15.0", "webdriver-manager>=4.0.0"]

[project.urls]
Repository = "https://github.com/CeasarSmj/comfyui_simplePromptManage"
#  Used by Comfy Registry https://registry.comfy.org
Documentation = "https://github.com/CeasarSmj/comfyui_simplePromptManage/wiki"
"Bug Tracker" = "https://github.com/CeasarSmj/comfyui_simplePromptManage/issues"

[tool.comfy]
PublisherId = "ceasarsmj"
DisplayName = "comfyui_simplePromptManage"
Icon = "web/logo.png"
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility

